Alternative approach to module alias using tsconfig paths.
Instead of using pnpm, alias the module using tsconfig paths.
    
```
"paths": {
  "~react-native": ["./node_modules/react-native-web"],
  "~/*": ["./app/*"],
},
```
    
Remix currently only accepts paths with prefixes with ~ or ., as mentioned in this [this issues](https://github.com/remix-run/remix/issues/1405), but this has the added benefit of allowing us to import react-native as well, or dependencies that have react-native as a peer dependency.